First checkin!
This is a first run through of todo-py. Hoping to get it up to an
awesome state in the near future. Made a simple plugin interface using
some haxors I found online, and an abuse of argparse. woot. As far as
what you can and cannot do right now, you can add and remove entries,
and you can use birdseye. That's about it.

Signed-off-by: Jeff Mickey <j@codemac.net>