Fixed project/context printing.
This change really shows the need for classes and other abstractions for
the todofile/todoentry formats. There is a clear format, and that needs
to be described in a single place, and formatting/printing needs to be
in a single place as well. There is way too much boilerplate code right
now.

Signed-off-by: Jeff Mickey <j@codemac.net>